直接在本文件及生成代码，controller没有修改完整
