+++ src/main/java/com/zjy/springbootadmin/SpringbootAdminApplication.java
@@ -1,7 +1,5 @@
 package com.zjy.springbootadmin;
 
-import com.zjy.springbootadmin.mapper.UserMapper;
-import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,9 +12,6 @@ import javax.annotation.Resource;
 @SpringBootApplication
 @EnableWebMvc
 public class SpringbootAdminApplication {
-
-    @Resource
-    private UserMapper userMapper;
 
     public static void main(String[] args) {
         SpringApplication.run(SpringbootAdminApplication.class, args);
