--- src/main/java/org/zjy/springbootadmin/entity/User.java
@@ -1,127 +1,0 @@
-package org.zjy.springbootadmin.entity;
-
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
-import java.io.Serializable;
-import java.time.LocalDateTime;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-
-/**
- * <p>
- * 
- * </p>
- *
- * @author zjy
- * @since 2022-08-15
- */
-@TableName("sys_user")
-@ApiModel(value = "User对象", description = "")
-public class User implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-      @ApiModelProperty("主键")
-        @TableId(value = "id", type = IdType.AUTO)
-      private Integer id;
-
-      @ApiModelProperty("用户名")
-      private String username;
-
-    private String password;
-
-      @ApiModelProperty("昵称")
-      private String nickname;
-
-      @ApiModelProperty("邮箱")
-      private String email;
-
-      @ApiModelProperty("电话")
-      private String phone;
-
-      @ApiModelProperty("地址")
-      private String address;
-
-      @ApiModelProperty("创建时间")
-      private LocalDateTime createTime;
-
-    
-    public Integer getId() {
-        return id;
-    }
-
-      public void setId(Integer id) {
-          this.id = id;
-      }
-    
-    public String getUsername() {
-        return username;
-    }
-
-      public void setUsername(String username) {
-          this.username = username;
-      }
-    
-    public String getPassword() {
-        return password;
-    }
-
-      public void setPassword(String password) {
-          this.password = password;
-      }
-    
-    public String getNickname() {
-        return nickname;
-    }
-
-      public void setNickname(String nickname) {
-          this.nickname = nickname;
-      }
-    
-    public String getEmail() {
-        return email;
-    }
-
-      public void setEmail(String email) {
-          this.email = email;
-      }
-    
-    public String getPhone() {
-        return phone;
-    }
-
-      public void setPhone(String phone) {
-          this.phone = phone;
-      }
-    
-    public String getAddress() {
-        return address;
-    }
-
-      public void setAddress(String address) {
-          this.address = address;
-      }
-    
-    public LocalDateTime getCreateTime() {
-        return createTime;
-    }
-
-      public void setCreateTime(LocalDateTime createTime) {
-          this.createTime = createTime;
-      }
-
-    @Override
-    public String toString() {
-        return "User{" +
-              "id=" + id +
-                  ", username=" + username +
-                  ", password=" + password +
-                  ", nickname=" + nickname +
-                  ", email=" + email +
-                  ", phone=" + phone +
-                  ", address=" + address +
-                  ", createTime=" + createTime +
-              "}";
-    }
-}
