--- src/main/java/org/zjy/springbootadmin/mapper/UserMapper.java
@@ -1,16 +1,0 @@
-package org.zjy.springbootadmin.mapper;
-
-import org.zjy.springbootadmin.entity.User;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-/**
- * <p>
- *  Mapper 接口
- * </p>
- *
- * @author zjy
- * @since 2022-08-15
- */
-public interface UserMapper extends BaseMapper<User> {
-
-}
