+++ src/main/resources/mapper/UserMapper.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.zjy.springbootadmin.mapper.UserMapper">
+<mapper namespace="com.zjy.springbootadmin.mapper.UserMapper">
 
 </mapper>
